Records: Use txt files to keep track of data.
We now started to use text files, and Perl scripts to turn those
text files into JavaScript files. That way, it's way easier to add
new data.

We also used more general classes, so we can later branch out
to different sports.

In this patch, we

    * Used a text file with flag data. This now falls directly under
      the Sports section, as it's non Speedskating specific.
    * Used a text file with rink data. Rinks will be instances of
      Venues; the latter defined under the Sports section. Later
      on, we may want to subclass Rinks as Venues.
    * Used a text file with skater data. Skaters will be instances
      of Athletes; the latter defined under the Sports section.
      Later on, we may want to subclass Skaters as Athletes.

To ease parsing, all "keys" we use to identify skaters and rinks,
will contain characters restricted to 'a-z' plus '-' and '_'.
Country codes will be three characters from the set 'A-Z' plus '1'.