Sports: Rewriting the progression.
We now keep track of the progression in plain text files. We lifted
the calculation of the improvement of a record, and its duration to
the Perl processing script.

We created (to be expanded) a "Record" class. An instance holds the
progression of the record, which, for now, is indentified by a sport
(say speedskating), a discipline (say 1000m), a gender (say female),
an age category (for now, we just have senior) and a type (for now,
we just have 'world', but this allows us to also track olympic, european
and national records).

The progression is basically a list (in date order) of records; each entry
is of the class Record_Entry (a special class Suspension is used to mark
suspensions of the record progression).