Trapped: Parse Betza
Just a subset though. We recognize the f, b, r, l, v, s, and h modifiers.
Also, the standard leapers, the generalized leaper ("(X, Y)"), the slider
(suffix 0), and the limited slider (suffix N > 0).

This is all put in a struct which list "parts", where parts leap in a
single direction, with the option to turn it into a slider.

For now, we do not recognize combined or chain movers. Nor croaked movers
or roses.