make_blogs: Handle svg-grid: plain.
This will act like an svg-grid: none, except if will throw in an inner
div with the "display: block" style. This will cause the content to
render as a normal block, but centered horizontally and vertically.

Perhaps all the svg-grid: none could be handled that way, but I do
not want to risk breaking something.