Trapped: Show compact heatmaps.
Using either the -c option, or a 'c' argument to the --heatmap option,
we show a compact heatmap.

A compact heatmap will not have any divider lines, use a single space
between values (so we may end up with lines of unequal size, and values
not vertically aligned), and use '.' for zero values.